Interaction work.
Interactions are semi functional.
Made an interface with 3 methods, for hoverEnter, hover exit, and interact button pressed.

If you want to make something interactable, give it a collider, set its layer to "Interactable" and give it a component that inherits from IInteractable.

If you want the outlines to work, add the "Outline" material as a 2nd material in the MeshRenderer.

Right now working on the panel in the tram. currently interacting with it moves the player to in front of the panel, looking at it. still thinking about how i want the actual ui interaction to work.