Interaction Progress 2
Fixed small issues
Make text in tram panels a world canvas with buttons

Bugs:
Figure out how to make camera's x rotation point to the target object. rn it is only lerping the y rotation and the players position, so if you are looking upwards, you can get the panel to be cut off.