Tram panels redone
somebody did them with ui buttons, but the cursor is disabled, and also they didn't place them as a child of a canvas, so they would not work. re did my old system for them, as well as set up a default position in the tram to load the player to.